moved BooleanConverter to TypeHelpers
